lille round system fix

+++ Common/src/main/java/data/componentdata/Health.java
@@ -35,7 +35,7 @@ public class Health
     }
 
 
-    public double getMaxHp()
+    public float getMaxHp()
     {
         return maxHp;
     }
